+++ examples/data-sources/uptimerobot_tag/data-source.tf
@@ -1,0 +1,7 @@
+data "uptimerobot_tag" "production" {
+  name = "production"
+}
+
+output "production_tag_id" {
+  value = data.uptimerobot_tag.production.id
+}