+++ internal/provider/provider.go
@@ -109,7 +109,16 @@ func (p *UptimeRobotProvider) Resources(ctx context.Context) []func() resource.R
 }
 
 func (p *UptimeRobotProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
-	return []func() datasource.DataSource{}
+	return []func() datasource.DataSource{
+		NewAllAlertContactsDataSource,
+		NewAlertContactDataSource,
+		NewAlertContactsDataSource,
+		NewCurrentUserDataSource,
+		NewIPRangesDataSource,
+		NewIntegrationDataSource,
+		NewTagDataSource,
+		NewTagsDataSource,
+	}
 }
 
 func (p *UptimeRobotProvider) Functions(ctx context.Context) []func() function.Function {